realupdatedrequirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
aioredis==2.0.1
aiosignal==1.3.1
albumentations==2.0.8
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
beautifulsoup4==4.13.3
blessed==1.20.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
colorful==0.5.6
cycler==0.11.0
distlib==0.3.9
enum34==1.1.10
exceptiongroup==1.2.2
filelock==3.12.2
flatbuffers==25.2.10
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
funcsigs==1.0.2
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
gorilla==0.4.0
gpustat==1.1.1
grpcio==1.62.3
hyperopt==0.2.7
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
matplotlib==3.5.3
msgpack==1.0.5
multidict==6.0.5
munch==4.0.0
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.570.86
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.2.0
pretrainedmodels==0.7.4
prometheus-client==0.17.1
proto-plus==1.26.1
protobuf==4.24.4
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.21
pyparsing==3.1.4
pyrsistent==0.19.3
pystopwatch2==0.1.1
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
ray==0.6.0
redis==3.4.1
requests==2.31.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setproctitle==1.3.3
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soupsieve==2.4.1
tensorboardX==2.6.2.2
theconf==0.1.0
threadpoolctl==3.1.0
timm==0.3.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.7
virtualenv==20.21.0
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git/@7021d63a49106e22c79b40564a7d39930e7b0f53
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
